Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_flex-shorthand-number-expected.txt


FAIL flexbox | computed style | flex: number assert_equals: expected "0px" but got "0%"
